+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v2.3.4
 * fix esm modules
+* fix browser build
+* fix hang
 
 ## v2.3.3
 * fix EmberTemplate is not a valid type of dependencies are not aligned